refactor : Refactor car racing
