+++ src/main/java/car/racing/CarRacingViewController.java
@@ -2,6 +2,8 @@ package car.racing;
 
 
 
+import car.racing.domain.CarRacingGame;
+import car.racing.domain.CarRacingWinners;
 import car.racing.view.ResultViewContract;
 
 import java.util.stream.IntStream;
