+++ src/main/java/car/racing/domain/CarRacingGame.java
@@ -1,17 +1,13 @@
-package car.racing;
+package car.racing.domain;
 
-import car.racing.domain.Car;
+import car.racing.Forwardable;
 
 import java.util.List;
-import java.util.Random;
 
 public class CarRacingGame {
 
-    private static final int LIMIT = 10;
-
     private final Forwardable forwardable;
     private final List<Car> cars;
-    private final Random random = new Random();
 
     public CarRacingGame(List<Car> cars, Forwardable forwardable) {
         this.cars = cars;
@@ -19,9 +15,7 @@ public class CarRacingGame {
     }
 
     public void forwardCarEachTry() {
-        cars.stream()
-                .filter(car -> forwardable.forwardable(random.nextInt(LIMIT)))
-                .forEach(Car::forward);
+        cars.forEach(car -> car.forward(forwardable));
     }
 
     public List<Car> getCars() {
