+++ src/test/java/car/racing/CarRacingViewControllerTest.java
@@ -1,6 +1,8 @@
 package car.racing;
 
 import car.racing.domain.Car;
+import car.racing.domain.CarRacingGame;
+import car.racing.domain.CarRacingWinners;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
